rmiloss update

+++ generator/dataset.py
@@ -72,7 +72,7 @@ class ImageDataset(torch.utils.data.Dataset):
         #lab = lab.squeeze()
         name = image_path.split("/")[-1].split(".")[0]
         try:
-            dist_map = np.load(label_path.split(".")[0] + ".npy").squeeze()
+            dist_map = np.int16(np.load(label_path.split(".")[0] + ".npy").squeeze())
         except Exception as e:
             print(e)
             dist_map = []
